Add usage and llm calls methods
